refactor: generator package 이동
- number 패키지 아래 위치하도록 수정
